+++ libsolidity/analysis/SyntaxChecker.h
@@ -83,6 +83,7 @@ private:
 	bool visit(PlaceholderStatement const& _placeholderStatement) override;
 
 	bool visit(ContractDefinition const& _contract) override;
+	void endVisit(ContractDefinition const& _contract) override;
 	bool visit(FunctionDefinition const& _function) override;
 	bool visit(FunctionTypeName const& _node) override;
 
@@ -102,7 +103,7 @@ private:
 	bool m_versionPragmaFound = false;
 
 	int m_inLoopDepth = 0;
-	bool m_isInterface = false;
+	std::optional<ContractKind> m_currentContractKind;
 
 	SourceUnit const* m_sourceUnit = nullptr;
 };