+++ test/libsolidity/syntaxTests/freeFunctions/free_function_qualified_modifier.sol
@@ -1,0 +1,9 @@
+contract C {
+  modifier someModifier() { _; }
+}
+
+function fun() C.someModifier {
+
+}
+// ----
+// ParserError 2314: (65-66): Expected '{' but got '.'