+++ test/libsolidity/syntaxTests/freeFunctions/free_function_shadowing.sol
@@ -1,0 +1,9 @@
+function f() {}
+contract C {
+  function f() public {}
+  function g() public {
+    f();
+  }
+}
+// ----
+// Warning 2519: (31-53): This declaration shadows an existing declaration.