+++ test/libsolidity/syntaxTests/freeFunctions/free_storage.sol
@@ -1,0 +1,4 @@
+struct S { uint x; }
+function fun(S storage) {
+}
+// ----