+++ test/libsolidity/syntaxTests/freeFunctions/struct_after_function.sol
@@ -1,0 +1,3 @@
+function f(S storage g) view returns (uint) { S storage t = g; return t.x; }
+struct S { uint x; }
+// ----